refactor(gen): split into smaller files
