+++ src/constants.py
@@ -5,7 +5,7 @@ CONFIG = get_yaml()
 
 WIDTH = CONFIG['gen_width']
 HEIGHT = CONFIG['gen_height']
-STEP = CONFIG['gen_step']
+STEP = CONFIG['gen_steps']
 API = CONFIG['gen_api']
 FOLDER = CONFIG['gen_folder']
 URL = CONFIG['gen_url']
